requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cairocffi==1.7.1
cffi==1.17.1
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.2
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
celery==5.4.0
redis==5.2.1
google-genai
requests
python-dotenv
cinemagoer
beautifulsoup4